+++ gradle.properties
@@ -1,3 +1,3 @@
-version=7.2.5
+version=7.2.6-SNAPSHOT
 group=io.muoncore
 exclude=doc,muon-examples
