+++ muon-examples/src/main/java/io/muoncore/examples/ServicePublishHotStream.java
@@ -1,4 +1,4 @@
-package com.simplicity.services;
+package io.muoncore.examples;
 
 import io.muoncore.Muon;
 import io.muoncore.MuonBuilder;
