+++ muon-examples/src/main/java/io/muoncore/examples/ServiceStreamConsumer.java
@@ -1,23 +1,26 @@
-package com.simplicity.services;
+package io.muoncore.examples;
 
 import io.muoncore.Muon;
 import io.muoncore.MuonBuilder;
 import io.muoncore.config.AutoConfiguration;
 import io.muoncore.config.MuonConfigBuilder;
+import io.muoncore.protocol.reactivestream.client.StreamData;
+import reactor.rx.broadcast.Broadcaster;
 
 import java.io.IOException;
+import java.net.URI;
 import java.net.URISyntaxException;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 
-import static io.muoncore.protocol.requestresponse.server.HandlerPredicates.all;
-
-public class ServiceAsapConsumer {
+public class ServiceStreamConsumer {
 
     public static void main(String[] args) throws URISyntaxException, InterruptedException, NoSuchAlgorithmException, KeyManagementException, IOException {
 
+        String serviceName = "myclient";
+
         AutoConfiguration config = MuonConfigBuilder
-                .withServiceIdentifier("awesomeService")
+                .withServiceIdentifier(serviceName)
                 .withTags("node", "awesome")
                 .build();
 
@@ -25,8 +28,11 @@ public class ServiceAsapConsumer {
 
         muon.getDiscovery().blockUntilReady();
 
-        muon.handleRequest(all(), response -> {
-            response.ok("Hellow");
+        Broadcaster<StreamData> b = Broadcaster.create();
+        b.consume(s -> {
+            System.out.println("DATA: " + s);
         });
+
+        muon.subscribe(new URI("stream://awesomeservicequery/ticktock"), b);
     }
 }
