+++ muon-examples/src/test/groovy/io/muoncore/examples/EventIntegrationSpec.groovy
@@ -1,4 +1,4 @@
-package com.simplicityitself
+package io.muoncore.examples
 
 import com.google.common.eventbus.EventBus
 import io.muoncore.Muon
